Almost complete, please look at this Description
I don't know how to show this calendar look like our  calendar on PC. IF
the first day of the month on Sunday then calendar looks like this.
However if the first day of month is on Wednesday , then we have to edit
our GridLayout(6,7). If it is on different day, we may edit our
GridLayout again. I didn't  embed the data yet.